Moving items from AppHeaderDropdown into shared
